requirements.txt
matplotlib
h5py
numpy
imageio
selenium
requests
pillow